Workbranch

+++ lib/graphics.h
@@ -19,7 +19,7 @@
 #include <iostream>
 #include <fstream>
 
-#include "ppm_image.h"
+#include "./ppm_image.h"
 
 namespace hxtk {
 namespace graphics {
@@ -62,9 +62,9 @@ class Graphics {
   void init(int, int);
   bool pointWithinPolygon(graphics::Point, graphics::Polygon);
   bool rayCrossesEdge(graphics::Point, graphics::Point, graphics::Point);
-    
+
   constexpr static double kEpsilon = std::numeric_limits<double>::min();
-  constexpr static int kNumLevels = 255;
+  constexpr static uint8_t kNumLevels = 255;
   std::vector<graphics::Polygon> polygons_;
   std::vector<uint8_t> canvas_;
   int width_ = -1;
